added and changed scripts
changed the character movement from using transform to using AddForce().
added tags to the traps and characters
made a Traps PreFab with the corisponding scripts.
from player side: when hitting the wrong trap the player will get booped back, if the player hits the correct trap it disables it.
form the trap side: if the player hits the wrong trap, 5 seconds are taken off the timer.
if player hits the correct trap, nothing happens.
also added a timer and a timer script.